oo-ratchet: split writer into apply.py so the tool passes its own OO check (vox-wbx1)
The reconcile rename-awareness work pushed writer.py to 312 lines, over the
300 module_size threshold -- a PY-BS-5 self-check failure that make check
missed because check-oo scores src/punt_vox/, not tools/. Extract the
plan-application mechanics (apply/write-or-refuse/prune/report + the
regressed/deltas metric helpers) into a cohesive PlanApplier in apply.py;
BaselineWriter keeps scope resolution, the fail-closed contract, and
dispatch. writer 178, apply 157, both under threshold; behavior unchanged.
